Adding ability to change sequence input mode
